requirements.txt
crewai
firecrawl-py 
crewai[tools]
streamlit
groq
python-dotenv